+++ pkg/clusters/types/kind/utils.go
@@ -9,9 +9,11 @@ import (
 	"os/exec"
 	"sync"
 
+	"gopkg.in/yaml.v2"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/kind/pkg/apis/config/v1alpha4"
 
 	"github.com/kong/kubernetes-testing-framework/pkg/clusters"
 )
@@ -34,6 +36,14 @@ func NewFromExisting(name string) (clusters.Cluster, error) {
 		addons: make(clusters.Addons),
 	}, nil
 }
+
+// -----------------------------------------------------------------------------
+// Private Consts & Vars
+// -----------------------------------------------------------------------------
+
+const (
+	defaultCalicoManifests = "https://projectcalico.docs.tigera.io/manifests/calico.yaml"
+)
 
 // -----------------------------------------------------------------------------
 // Private Functions - Cluster Management
@@ -89,4 +99,54 @@ func clientForCluster(name string) (*rest.Config, *kubernetes.Clientset, error) 
 
 	clientset, err := kubernetes.NewForConfig(cfg)
 	return cfg, clientset, err
+}
+
+const defaultKindConfig = `---
+kind: Cluster
+apiVersion: kind.x-k8s.io/v1alpha4
+`
+
+func (b *Builder) ensureConfigFile() error {
+	if b.configPath == nil {
+		f, err := os.CreateTemp(os.TempDir(), "ktf-kind-config")
+		if err != nil {
+			return err
+		}
+		defer f.Close()
+
+		_, err = f.WriteString(defaultKindConfig)
+		if err != nil {
+			return err
+		}
+
+		filename := f.Name()
+		b.configPath = &filename
+	}
+
+	return nil
+}
+
+func (b *Builder) disableDefaultCNI() error {
+	if err := b.ensureConfigFile(); err != nil {
+		return err
+	}
+
+	configYAML, err := os.ReadFile(*b.configPath)
+	if err != nil {
+		return err
+	}
+
+	kindConfig := v1alpha4.Cluster{}
+	if err := yaml.Unmarshal(configYAML, &kindConfig); err != nil {
+		return err
+	}
+
+	kindConfig.Networking.DisableDefaultCNI = true
+
+	configYAML, err = yaml.Marshal(kindConfig)
+	if err != nil {
+		return err
+	}
+
+	return os.WriteFile(*b.configPath, configYAML, 0600) //nolint:gomnd
 }